@@ -2,12 +2,12 @@ name: flutter_ecommerce
 description: "A new Flutter application."
 
 
-publish_to: 'none' 
+publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: "3.3.3"
+  sdk: "3.4.0-282.2.beta"
 
 
 dependencies:
@@ -16,7 +16,7 @@ dependencies:
 
   # Utility packages
   http: ^1.2.1
-  intl: ^0.18.0
+  intl: ^0.19.0
   logger: ^2.1.0
   lottie: ^2.7.0
   shimmer: ^3.0.0
@@ -26,14 +26,11 @@ dependencies:
   pdf: ^3.1.0
   flutter_pdfview: ^1.3.2
   path_provider: ^2.0.14
-  dio: ^5.4.3+1
-  flutter_polyline_points: ^2.0.0
-  google_maps_flutter: ^2.6.1
-  dash_chat_2: ^0.0.18
-  chat_gpt_sdk: ^2.2.5
   connectivity_plus: ^5.0.1
+  carousel_slider: ^4.2.1
   flutter_native_splash: ^2.3.2
   smooth_page_indicator: ^1.1.0
+  readmore: ^3.0.0
 
   # Icons
   iconsax: ^0.0.8
@@ -44,13 +41,6 @@ dependencies:
   get_storage: ^2.1.1
   firebase_core: ^2.29.0
   firebase_storage: ^11.5.4
-
-  # Product Specific
-  readmore: ^2.2.0
-  carousel_slider: ^4.2.1
-
-
-
 
   # Firebase
   firebase_auth: ^4.19.2
