@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_ecommerce/common/widgets/custom_shapes/containers/rounded_container.dart';
 import 'package:flutter_ecommerce/common/widgets/images/t_rounded_image.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/product_title_text.dart';
-import 'package:flutter_ecommerce/features/shop/controllers/product_controller.dart';
+import 'package:flutter_ecommerce/features/shop/controllers/product/product_controller.dart';
 import 'package:flutter_ecommerce/features/shop/models/product_model.dart';
 import 'package:flutter_ecommerce/features/shop/screens/product%20detail/product_detail.dart';
 import 'package:flutter_ecommerce/utils/constants/colors.dart';
@@ -30,7 +30,7 @@ class TProductCardVertical extends StatelessWidget {
     final dark = THelperFunctions.isDarkMode(context);
 
     return GestureDetector(
-      onTap: () => Get.to (()=> const ProductDetail()),
+      onTap: () => Get.to(() => ProductDetail(product: product,)),
       child: Container(
         width: 180,
         padding: const EdgeInsets.all(1),
